#8 - 도메인 클래스 설계 (게시글, 댓글) 도메인 정보를 설계해 보았다. 기초적인 게시글과 댓글 표현이 가능하도록 함 DB 와 연동을 위한 추가적인 표현은 추후에 업데이트 될 수 있음
