--- src/main/java/com/fastcampus/fastcampusprojectboard/config/JpaConfig.java
@@ -1,19 +1,0 @@
-package com.fastcampus.fastcampusprojectboard.config;
-
-import org.springframework.context.annotation.Bean;
-import org.springframework.context.annotation.Configuration;
-import org.springframework.data.domain.AuditorAware;
-import org.springframework.data.jpa.repository.config.EnableJpaAuditing;
-
-import java.util.Optional;
-
-@EnableJpaAuditing
-@Configuration
-public class JpaConfig {
-
-    @Bean
-    public AuditorAware<String> auditorAware(){
-        return() -> Optional.of("uno");
-    }
-
-}
