--- src/main/java/com/fastcampus/fastcampusprojectboard/config/ThyleafConfig.java
@@ -1,36 +1,0 @@
-package com.fastcampus.fastcampusprojectboard.config;
-
-
-import lombok.Getter;
-import lombok.RequiredArgsConstructor;
-import org.springframework.boot.context.properties.ConfigurationProperties;
-import org.springframework.boot.context.properties.ConstructorBinding;
-import org.springframework.context.annotation.Bean;
-import org.springframework.context.annotation.Configuration;
-import org.thymeleaf.spring5.templateresolver.SpringResourceTemplateResolver;
-
-@Configuration
-public class ThyleafConfig {
-
-    @Bean
-    public SpringResourceTemplateResolver thymeleafTemplateResolver(
-            SpringResourceTemplateResolver defaultTemplateResolver,
-            Thymeleaf3Properties thymeleaf3Properties
-    ){
-        defaultTemplateResolver.setUseDecoupledLogic(thymeleaf3Properties.isDecoupledLogic());
-
-        return defaultTemplateResolver;
-    }
-
-    @RequiredArgsConstructor
-    @Getter
-    @ConstructorBinding
-    @ConfigurationProperties("spring.thymeleaf3")
-    public static class Thymeleaf3Properties{
-        /**
-         * Use Thymeleaf 3 Decoupled Logic
-         */
-        private  final boolean decoupledLogic;
-
-    }
-}
