--- src/main/java/com/fastcampus/fastcampusprojectboard/domain/Article.java
@@ -1,70 +1,0 @@
-package com.fastcampus.fastcampusprojectboard.domain;
-
-import lombok.EqualsAndHashCode;
-import lombok.Getter;
-import lombok.Setter;
-import lombok.ToString;
-import org.springframework.data.annotation.CreatedBy;
-import org.springframework.data.annotation.CreatedDate;
-import org.springframework.data.annotation.LastModifiedBy;
-import org.springframework.data.annotation.LastModifiedDate;
-
-import javax.persistence.*;
-import java.time.LocalDateTime;
-import java.util.LinkedHashSet;
-import java.util.Objects;
-import java.util.Set;
-
-@Getter
-@ToString
-@Table(indexes = {
-        @Index(columnList = "title"),
-        @Index(columnList = "hashtag"),
-        @Index(columnList = "createdAt"),
-        @Index(columnList = "createdBy"),
-})
-@Entity
-public class Article {
-
-    @Id
-    @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private Long id;
-
-    @Setter @Column(nullable = false) private String title;   // 제목
-    @Setter @Column(nullable = false, length = 10000) private String content; // 본문
-    @Setter private String hashtag; // 해시태그
-
-    @OrderBy("id")
-    @OneToMany(mappedBy = "article", cascade = CascadeType.ALL)
-    @ToString.Exclude
-    private final Set<ArticleComment> articleComments = new LinkedHashSet<>();
-
-    @CreatedDate @Column(nullable = false) private LocalDateTime createdAt; // 생성일시
-    @CreatedBy @Column(nullable = false, length = 100) private String createdBy; // 생성자
-    @LastModifiedDate @Column(nullable = false) private LocalDateTime modifiedAt; // 수정일시
-    @LastModifiedBy private String modifiedBy; // 수정자
-
-    protected Article(){}
-
-    private Article(String title, String content, String hashtag) {
-        this.title = title;
-        this.content = content;
-        this.hashtag = hashtag;
-    }
-
-    public static Article of(String title, String content, String hashtag) {
-        return new Article(title, content, hashtag);
-    }
-
-    @Override
-    public boolean equals(Object o) {
-        if (this == o) return true;
-        if (!(o instanceof Article article)) return false;
-        return id != null && id.equals(article.id);
-    }
-
-    @Override
-    public int hashCode() {
-        return Objects.hash(id);
-    }
-}
