--- src/main/java/com/fastcampus/fastcampusprojectboard/domain/ArticleComment.java
@@ -1,42 +1,0 @@
-package com.fastcampus.fastcampusprojectboard.domain;
-
-import lombok.EqualsAndHashCode;
-import lombok.Getter;
-import lombok.Setter;
-import lombok.ToString;
-import org.springframework.data.annotation.CreatedBy;
-import org.springframework.data.annotation.CreatedDate;
-import org.springframework.data.annotation.LastModifiedBy;
-import org.springframework.data.annotation.LastModifiedDate;
-
-import javax.persistence.*;
-import java.time.LocalDateTime;
-
-@Getter
-@EqualsAndHashCode
-@ToString
-@Table(indexes = {
-        @Index(columnList = "title"),
-        @Index(columnList = "hashtag"),
-        @Index(columnList = "createdAt"),
-        @Index(columnList = "createdBy"),
-})
-@Entity
-public class ArticleComment {
-
-    @Id
-    @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private Long id;
-
-    @Setter @ManyToOne(optional = false) private Article article;
-    @Setter @Column(nullable = false, length = 500) private String content;
-
-    @CreatedDate
-    @Column(nullable = false) private LocalDateTime createdAt; // 생성일시
-    @CreatedBy
-    @Column(nullable = false, length = 100) private String createdBy; // 생성자
-    @LastModifiedDate
-    @Column(nullable = false) private LocalDateTime modifiedAt; // 수정일시
-    @LastModifiedBy
-    private String modifiedBy; // 수정자
-}
