--- src/main/java/com/fastcampus/fastcampusprojectboard/repository/ArticleRepository.java
@@ -1,28 +1,0 @@
-package com.fastcampus.fastcampusprojectboard.repository;
-
-import com.fastcampus.fastcampusprojectboard.domain.Article;
-import com.fastcampus.fastcampusprojectboard.domain.QArticle;
-import com.querydsl.core.types.dsl.DateTimeExpression;
-import com.querydsl.core.types.dsl.StringExpression;
-import org.springframework.data.jpa.repository.JpaRepository;
-import org.springframework.data.querydsl.QuerydslPredicateExecutor;
-import org.springframework.data.querydsl.binding.QuerydslBinderCustomizer;
-import org.springframework.data.querydsl.binding.QuerydslBindings;
-import org.springframework.data.rest.core.annotation.RepositoryRestResource;
-
-@RepositoryRestResource
-public interface ArticleRepository extends
-        JpaRepository<Article, Long>,
-        QuerydslPredicateExecutor<Article>,
-        QuerydslBinderCustomizer<QArticle>
-{
-    @Override
-    default void customize(QuerydslBindings bindings, QArticle root){
-        bindings.excludeUnlistedProperties(true);
-        bindings.including(root.title, root.content, root.hashtag, root.createdAt, root.createdBy);
-        bindings.bind(root.content).first(StringExpression::containsIgnoreCase);
-        bindings.bind(root.hashtag).first(StringExpression::containsIgnoreCase);
-        bindings.bind(root.createdAt).first(DateTimeExpression::eq);
-        bindings.bind(root.createdBy).first(StringExpression::containsIgnoreCase);
-    }
-}
