+++ src/main/java/com/fastcampus/projectboard/FastcampusProjectBoardApplication.java
@@ -1,4 +1,4 @@
-package com.fastcampus.fastcampusprojectboard;
+package com.fastcampus.projectboard;
 
 import org.springframework.boot.SpringApplication;
 import org.springframework.boot.autoconfigure.SpringBootApplication;
