+++ src/main/java/com/fastcampus/projectboard/domain/ArticleComment.java
@@ -1,0 +1,14 @@
+package com.fastcampus.projectboard.domain;
+
+import java.time.LocalDateTime;
+
+public class ArticleComment {
+    private Long id;
+    private Article article; // 게시글 (ID)
+    private String content; // 본문
+
+    private LocalDateTime createdAt;
+    private String createBy;
+    private LocalDateTime modifiedAt;
+    private String modifiedBy;
+}
