+++ src/test/java/com/fastcampus/projectboard/FastcampusProjectBoardApplicationTests.java
@@ -1,4 +1,4 @@
-package com.fastcampus.fastcampusprojectboard;
+package com.fastcampus.projectboard;
 
 import org.junit.jupiter.api.Test;
 import org.springframework.boot.test.context.SpringBootTest;
