+++ src/test/java/com/fastcampus/projectboard/controller/ArticleControllerTest.java
@@ -1,4 +1,4 @@
-package com.fastcampus.fastcampusprojectboard.controller;
+package com.fastcampus.projectboard.controller;
 
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
@@ -7,8 +7,6 @@ import org.springframework.http.MediaType;
 import org.springframework.test.web.servlet.MockMvc;
 import org.springframework.test.web.servlet.request.MockMvcRequestBuilders;
 import org.springframework.test.web.servlet.result.MockMvcResultMatchers;
-
-import static org.junit.jupiter.api.Assertions.*;
 
 @DisplayName("View 컨트롤러 - 게시글")
 @WebMvcTest
